+++ x/perp/v2/keeper/clearing_house.go
@@ -98,7 +98,10 @@ func (k Keeper) OpenPosition(
 		}
 	}
 
-	if err = k.afterPositionUpdate(ctx, market, *updatedAMM, traderAddr, *positionResp); err != nil {
+	changeType := "open_position"
+	if err = k.afterPositionUpdate(
+		ctx, market, *updatedAMM, traderAddr, *positionResp, changeType,
+	); err != nil {
 		return nil, err
 	}
 
@@ -527,6 +530,7 @@ func (k Keeper) afterPositionUpdate(
 	amm types.AMM,
 	traderAddr sdk.AccAddress,
 	positionResp types.PositionResp,
+	changeType string,
 ) (err error) {
 	// check bad debt
 	if !positionResp.BadDebt.IsZero() {
@@ -588,6 +592,8 @@ func (k Keeper) afterPositionUpdate(
 			BadDebt:          sdk.NewCoin(market.Pair.QuoteDenom(), positionResp.BadDebt.RoundInt()),
 			FundingPayment:   positionResp.FundingPayment,
 			BlockHeight:      ctx.BlockHeight(),
+			ExchangedMargin:  marginToVault.Neg().Sub(transferredFee),
+			ChangeType:       changeType,
 		},
 	)
 
@@ -761,12 +767,14 @@ func (k Keeper) ClosePosition(ctx sdk.Context, pair asset.Pair, traderAddr sdk.A
 		return positionResp, nil
 	}
 
+	changeType := "close_position"
 	if err = k.afterPositionUpdate(
 		ctx,
 		market,
 		*updatedAMM,
 		traderAddr,
 		*positionResp,
+		changeType,
 	); err != nil {
 		return nil, err
 	}